requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Image handling
pillow==10.4.0
opencv-python-headless==4.10.0.84

# Machine learning / HuggingFace
torch==2.4.1
transformers==4.44.2
huggingface-hub==0.25.1

# File uploads
python-multipart==0.0.9

# Data validation
pydantic==2.9.2
